BND cable tapping http://www.spiegel.de/spiegel/vorab/bnd-laesst-sich-abhoeren-von-verbindungen-deutscher-provider-genehmigen-a-926221.html#
Evidence that spying 2.0 is directly on the cables (APEX) in order to re-route/duplicate or compromise the traffic. Frankfurt/Main is one of the biggest Internet Nodes (there 5 on the entire world which the traffic needs to go trough) and a 6 one is build (China).

This shows that no OS/Software needs to be compromised and proves it when they simply can grab everything on the nodes directly, which is under the law now and 'legal'. Fucking ridiculous if you ask me. So basically only TLS 1.3 is the chance someone has (if that is not already cracked [they had 4 years for the draft and time to build strats to crack it].

Signed-off-by: CHEF-KOCH <CHEF-KOCH@protonmail.com>