Under section 702 FAA, only data that are associated with a specific selector are stored. For Upstream collection, this means only the communications that remain after the filtering process. These are processed (via decoded, formatted, etc.) and stored in NSA databases for a maximum of only 2 years. "Officially" (I found no evidence that this 2 years is 'controlled' somehow which means I doubt that it's only stored 2 years [confirmation needed].
Downstream collection under the PRISM program results in all the data associated with specific selectors that the big internet companies hand over to the FBI, which then forwards them to NSA. These are also processed and then stored for a maximum of 5 years [confirmation needed].

Data from FAA collection are usually stored in separate database partitions and are protected by the Exceptionally Controlled Information (ECI) compartment RAGTIME (RGT). Only analysts who are cleared for RAGTIME, have the specific need-to-know and who are authorized by the data owner have access to these data.

The information are provided in from here:
http://www.washingtonian.com/2013/02/27/ragtime-codename-of-nsas-secret-domestic-intelligence-program-revealed-in-new-book/

& from the paper:
- RAGTIME-A: foreign-to-foreign counter-terrorism (CT) data
- RAGTIME-B: data from foreign governments (FG) transiting the US
- RAGTIME-C: data related to counter-proliferation (CP) activities
- RAGTIME-P: domestic bulk collection of internet metadata

Signed-off-by: CHEF-KOCH <CHEF-KOCH@protonmail.com>