Google's Toronto Spying Idea Plan
"Sidewalk Toronto Data Governance Working Group" is known to be "cooperative" helping the gov. with "new ideas". However, Google is not directly involved here but the idea and the concept is based on their military "experience" and they used the app data from legit services (I assume Google Play Store [it's not mentioned]).

It's a full version which is untouched.